761: * adjust the size of the column * move type to the libs section
