+++ apps/frontend/src/components/ui/Table/TableActionMenu.tsx
@@ -19,7 +19,7 @@ import cn from '@libs/common/utils/className';
 import DropdownMenuItemType from '@libs/ui/types/dropdownMenuItemType';
 import DropdownMenu from '@/components/shared/DropdownMenu';
 import { Button } from '@/components/shared/Button';
-import TableAction from './TableAction';
+import TableAction from '../../../../../../libs/src/common/types/TableAction';
 
 interface TableActionMenuProps<TData, TValue> {
   actions: TableAction<TData, TValue>[];
